add poweruser concerns to the application config
